# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API document for UAMS backend server",
        "title": "Backend API",
        "contact": {
            "name": "Mr. Khai"
        },
        "version": "1.0"
    },
    "host": "http://iot.hcmue.space:8079",
    "basePath": "/v1",
    "paths": {
        "/v1/area": {
            "post": {
                "description": "Create area",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Area",
                "parameters": [
                    {
                        "description": "Fields need to create a area",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwagCreateArea"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Area"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete area using \"id\" field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Area By ID",
                "parameters": [
                    {
                        "description": "Area ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update area, must have \"id\" field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update Area By ID",
                "parameters": [
                    {
                        "description": "Fields need to update a area",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwagUpdateArea"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/area/{id}": {
            "get": {
                "description": "find area info by area id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Area By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Area ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Area"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/areas": {
            "get": {
                "description": "find all areas info",
                "produces": [
                    "application/json"
                ],
                "summary": "Find All Area",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Area"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway": {
            "delete": {
                "description": "Delete gateway using \"id\" field. Send deleted info to MQTT broker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Gateway By Gateway ID",
                "parameters": [
                    {
                        "description": "Gateway ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "type": "object"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "gateway_id": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update gateway, must have \"gateway_id\" field. Send updated info to MQTT broker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update Gateway By Gateway ID",
                "parameters": [
                    {
                        "description": "Fields need to update a gateway",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwagUpateGateway"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway/gateway_id/{gateway_id}": {
            "get": {
                "description": "find gateway and uhf info by gateway_id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Gateway By Gateway ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gateway_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Gateway"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway/{id}": {
            "get": {
                "description": "find gateway and uhf info by id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Gateway By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Gateway"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway_logs": {
            "get": {
                "description": "find all gateway logs info",
                "produces": [
                    "application/json"
                ],
                "summary": "Find All GatewayLog",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.GatewayLog"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway_logs/gateway_id/{gateway_id}/period/{from}/{to}": {
            "get": {
                "description": "find Gateway logs by period of time",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Gateway logs by period of time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GatewayLog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From Unix time",
                        "name": "from",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To Unix time",
                        "name": "to",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GatewayLog"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway_logs/gateway_id/{id}": {
            "get": {
                "description": "find gateway log info by Gateway id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find GatewayLog By Gateway ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GatewayLog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GatewayLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway_logs/period/{from}/{to}": {
            "get": {
                "description": "find Gateway logs by period of time",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Gateway logs by period of time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GatewayLog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From Unix time",
                        "name": "from",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To Unix time",
                        "name": "to",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GatewayLog"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateway_logs/{id}": {
            "get": {
                "description": "find gateway log info by id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find GatewayLog By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GatewayLog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GatewayLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/gateways": {
            "get": {
                "description": "find all gateways info",
                "produces": [
                    "application/json"
                ],
                "summary": "Find All Gateway",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Gateway"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/operation_logs": {
            "get": {
                "description": "find all operation logs info",
                "produces": [
                    "application/json"
                ],
                "summary": "Find All OperationLog",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.OperationLog"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/operation_logs/gateway_id/{gateway_id}": {
            "get": {
                "description": "find operation log info by gateway_id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Operation By GatewayID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GatewayLog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OperationLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/operation_logs/gateway_id/{gateway_id}/period/{from}/{to}": {
            "get": {
                "description": "find operation logs by period of time",
                "produces": [
                    "application/json"
                ],
                "summary": "Find operation logs by period of time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From Unix time",
                        "name": "from",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To Unix time",
                        "name": "to",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OperationLog"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/operation_logs/period/{from}/{to}": {
            "get": {
                "description": "find Operation logs by period of time",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Operation logs by period of time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From Unix time",
                        "name": "from",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "To Unix time",
                        "name": "to",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OperationLog"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/operation_logs/{id}": {
            "get": {
                "description": "find operation log info by id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Operation By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OperationLog"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/package_accesses": {
            "get": {
                "description": "find all package access log",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Package Access log",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PackageAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/package_accesses/area_id/{area_id}": {
            "get": {
                "description": "find all package access log by  and AreaID",
                "produces": [
                    "application/json"
                ],
                "summary": "Find Package Access log by AreaID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.PackageAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Area": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "manager": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.SwagCreateArea": {
            "type": "object",
            "properties": {
                "manager": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.SwagUpateGateway": {
            "type": "object",
            "properties": {
                "areaId": {
                    "type": "string"
                },
                "gatewayId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.SwagUpdateArea": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "manager": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorMsg": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    }
}